requirements.txt
asgiref==3.8.1
crispy-bootstrap5==2024.2
Django==5.0.6
django-crispy-forms==2.2
django-nested-admin==4.0.2
python-monkey-business==1.0.0
setuptools==70.1.1
six==1.16.0
sqlparse==0.5.0
